refactor : validator 클래스 제거. 유효성 검증은 각각 Players 와 Height 클래스에 위임
