+++ src/main/java/domain/Height.java
@@ -1,15 +1,13 @@
 package domain;
 
-import laddervalidate.HeightValidator;
-
 public class Height {
     private static final int FINISH_NUMBER = 0;
+    public static final String NOT_POSITIVE_ERROR_MESSAGE = "양의 정수만 입력해주세요.";
 
     private int height;
 
     public Height(String height) {
-        HeightValidator heightValidator = new HeightValidator();
-        heightValidator.validate(height);
+        validate(height);
         this.height = Integer.parseInt(height);
     }
 
@@ -19,5 +17,11 @@ public class Height {
 
     public void minusHeight() {
         this.height--;
+    }
+
+    private void validate(String height) {
+        if (Integer.parseInt(height) <= 0) {
+            throw new IllegalArgumentException(NOT_POSITIVE_ERROR_MESSAGE);
+        }
     }
 }
